working with stabilized curl formulation
